Refactor the models to use Double instead of Int for numbers with fraction in API call

+++ app/src/main/java/com/dynamo/spacex/data/network/model/pastlaunch/OrbitParams.kt
@@ -1,16 +1,16 @@
 package com.dynamo.spacex.data.network.model.pastlaunch
 
 data class OrbitParams(
-    val apoapsis_km: Int, // 500
+    val apoapsis_km: Double, // 500
     val arg_of_pericenter: Any?, // null
     val eccentricity: Any?, // null
     val epoch: Any?, // null
-    val inclination_deg: Int, // 39
+    val inclination_deg: Double, // 39
     val lifespan_years: Any?, // null
     val longitude: Any?, // null
     val mean_anomaly: Any?, // null
     val mean_motion: Any?, // null
-    val periapsis_km: Int, // 400
+    val periapsis_km: Double, // 400
     val period_min: Any?, // null
     val raan: Any?, // null
     val reference_system: String, // geocentric
